Use the package name from context when accessing the app info object
Issue:
In the old implementation, resource package name and app package name
were used interchangeably. This is wrong as they are not necessarily
the same.

When apps get repackaged by aapt, the package name in the manifest
is changed but package names in the resource table remain unchanged.
IconCompat runs into crash when trying to use the resource package name
to get the application info because it is still the old package name.

Solution:
Always use app package name to get the application info / resource object.
Only use resource package name when calling Resource.getIdentifier(..)
Resource full name is put in mObj1 and app package name is put
in string1.

Also updated the build.gradle of the test app so that it recompiles
whenever the core library, where IconCompat lives, is updated.

Test: Run IconCompatTest.java.
Test: Modify the sample app to load an icon from GooglePackageInstaller
   which is repackaged by aapt. The app no longer crashes with the fix.

BUG: 155906259

Change-Id: Ie689dc114ed7598160b5e3900b0a5d2632496a1b