Move all focus code to androidx.ui.core.focus
After moving code from ui-framework and ui-platform to ui-core, The focus code is spread across the androidx.ui.core.focus and androidx.ui.focus. This CL moves all the focus code to the androidx.ui.core.focus package.

Also fixed an issue with tests that failed when we called requestFocus() on a view with no size.

Bug: 156209815
Test: Ran the demo app locally, and updated all the affected tests
Change-Id: I53fb4d24fce09750c7450f47c9a537adb851a793