Fix zIndex is ignored for components with semantics
Components like Button, FloatingActionButton or TopAppBar adds Semantics on top of Surface which adds extra layout layer for now so even if the Button has some zIndex it will be ignored as PassThroughLayout has 0 zIndex. I added temporary workaround so PassThroughLayout will use the zIndex of the first child instead. This will be removed together with removing PassThroughLayout.

Fixes: 156582908
Test: manually, new test
Change-Id: Ia033958535d6d29c09c2af622d78585558cc73a9