Update NavBackStackEntry args on nav with singleTop
When navigating with the singleTop option, if the call to navigate
includes new arguments those are currently, ignored by the entry on the
backstack and the original arguments are kept.

We need to replace the original arguments with the new ones on the
back stack entry to reflect the most recent status of the destination.

Test: add test in NavControllerTest
Bug: 156545508
Change-Id: Icf2c92f424f57f815487d52f37383b9b11b42fa4