Undeprecate semantics and add ability to affect current layout node
This CL adds 'applyToChildLayoutNode' mode to semantics modifier, which makes is behave like proper modifier and modify the node it's being put on.

When set to 'true', it causes the old behavior where a
modifier is supposed to be used with PassThroughLayout and actually
applies to a child non-semantics layout node.  When set to 'false',
it instead has the new more intuitive collapsing behavior into the
layout node that it modifies.  'false' is the new default and I
un-deprecated semanticsCore since it's reasonable to begin using
with this.

It also made the root node have a normal SemanticsModifier and
SemanticsWrapper like every other semantics node.

Relnote: Modifier.semantics has been undeprecated to allow usages for high level components.

Test: checkbox and scroller tests has been changed to work with Modifier.testTag to verify that this functionality works
Bug: 153198273
Fixes: 154023028

Change-Id: I4cfdc837d5ac2d240af5a5ac6b755aebf800af15