Implemention of scroll orientation locking.
This implements a first pass of scroll
orientation locking that is akin to the
logic found in Android.

It also introduces ScrollGestureDetector
which is the same as DragGestureDetector
except that it disambiguates with other
ScrollGestureDetectors so that a given
pointer is only ever used to drag in
one orientation or the other.

The api that does the orientation locking
itself is experimental (we want to improve
on it) so it is RequiresOptIn.

Bug: 150289741
Fixes: 147609848
Test: New tests added in this CL.
Test: ./gradlew ui:ui-core:test

Change-Id: Ia611c96733f9749d9cc52d5ffbaf29644caf2e73