Add reasonable slot defaults for Scrollers
This CL adds Column and Row as a default to Vertical and Horizontal scrollers. This potentially saves one layout node and one big tab shift in the code, and it's easy to opt out for any reason.

Relnote: VerticalScroller now provides Column out of the box. HorizontalScroller now provides Row out of the box.

Bug: 157020670
Test: Tests updated correspondingly
Change-Id: Ieca5d185b9f6e950a7175b9daa7a9a511a439da2