Fixed issue where alpha was not applied in all cases
If a SolidColor brush was used and had an
alpha value baked into the color, it previously
would not be applied. Updated Brush implementation
to apply alpha value as it can provide a more optimal
alpha algorithm depending on the implementation.
Updated configurePaint method in DrawScope to set the
alpha if no Brush is provided (ex. drawImage)

Change-Id: I586dfcd21147e18e9d9f1add5d06c12861e60005
Fixes: 157166009
Test: Added test to DrawScopeTest + re-ran compose tests