Rename MotionEventRecorder methods for clarity
Because it isn't implementing an interface (anymore), the method name
for recording an event can be chosen freely. `recordEvent` is a much
better name than `sendEvent`, as it is not sending anything, but
recording it.

I renamed the `clear` method to `disposeEvents` to make it clear that it
is freeing up resources and that it is therefore very important to call
it after you're done with it. `clear` doesn't tell that story.

Bug: 145593752
Test: ./gradlew ui:ui-test:cC
Change-Id: I87be5d9c1f99608cf8b53bd37ae302bb557b16b4