Prototype: allow swapping projects with artifacts
Gradle dependency resolution does not allow replacing project
dependencies with artifacts. This CL creates a function that is
available on subprojects where they can reference a project
dependency with  which will swap it with the
latest artifact if the project does not exists.

Even though this is quite a hack, if we want to consider moving
partial projects into playground, it might be a feasible option