Always find the lastIn Fragment
Due to how ordered transactions occur, we should
always look for the lastIn Fragment, even if it
occurs before the firstOut Fragment.

This has a side effect of sometimes there being
a lastIn Fragment but no firstOut Fragment
(such as is the case for the first add()), so
we ensure that we only compute the shared
element transition when there's both a
firstOut and lastIn Fragment.

Test: tests get slightly farther
BUG: 147749580
Change-Id: I5dc060b2b4ba6c200b772fd7a6dcfd251e8bdb25