Always run gmaven diff all task
This CL fixes an edge case where if gmaven contents ever matches local
build, it will fail the build as no zip file will be generated for the
all output but buildOnServer requires that file.

I think it is best to always create it to ensure all of the work happened,
even when the zip file is empty.

Bug: 158707025
Test: playground bOS works
Change-Id: I9e7b36f1712fff67d7f9d32efac6ff05804104a7