Never collect for Flow<PageEvent> twice
This CL Fixes a bug where the Multicast logic could possibly
restart the Flow<PageEvent> if upsteam closes and a new downstream
arrives before we get the new PagedData.

Since Flow<PageEvent> (PagedData.flow) cannot be collected twice and
we already keep the history of events cached, this CL ensures that
if Multicaster is restarted, we won't connect to the original flow.

Instead, we'll just complete that flow as an empty flow and let
downstream receive history of events, if any exists.

I'm not changing Multicaster because from its perspective, restarting
upstream makes sense if there is a new downstream after previous
upstream closes. This behavior is specifc to CachedPageEventFlow hence
handled there.
We run Multicaster with keepAlive so there is no case where the
collection from upstream would stop unless:
* upstream closes
* we close the CachedPageEventFlow (usually due to a new one)
* scope is cancelled

Bug: 158784811
Test: CachedPageEventFlowTest, muzei
Change-Id: I58bcc4b2bde88b1c78ed1a96cf227e068127b47e