Fix inflated fragment lifecycle
When inflating an Activity view that uses the Fragment tag, the Fragment
state moves to VIEW_CREATED while its parent FragmentManager is still in
the CREATED state. In the current moveToState world, this force
fragments back down to CREATED, which immediately destroys their view.

Since FragmentManager gets its state from the mHost and there is no
VIEW_CREATED state associated with the host. Therefore, when in this
situation, we should move the newState up to VIEW_CREATED and allow
computeExpectedState() to move it back down if needed.

Test: added fragment lifecycle test (49 failing with USE_STATE_MANAGER)
Fixes: 158791416
Change-Id: I95e0c2e74c3a1fef44f9a741888a6e82391d370c