Add testing APIs for positions.
This removes the need to use onPositioned callback to test positions of
layout nodes. It also uses boundsInRoot which is what we should use for
the assertions instead of globalBounds.

As an example and proof of concept I have rewritten tests in TabTest
that were failing previously on cut-outs and this fixes them.

Bug: 157657482
Test: Added
Change-Id: Ie7c83615b4765b0dd4332f392b0af0eca278aa8f