Mitigate the wrong text field scrolling
We need to call View#requestRectangleOnScreen not to be covered by
software keyboard, but it is not clear when this API is safely called.

After getting focus, if we call this API too early, this API sometimes
bring the rectangle to wrong place.
On the other hand, we don't have good timing of calling this API after
getting focus. So, as a workaround, call that API when software keyboard
size has changed. However, there is no public API of getting software
keyboard visibility before API 29. So, use OnGlobalLayoutListener as a
workaround.

Bug: 138449572
Test: Manually done on API 21, 28, 30
Change-Id: I089ed5dd780bdd9a1d5305e7e668f82a660fa9c5