Fix flaky focus-related tests
KeyEvents are only received when the owner has both window focus and view focus. The flake was caused when the test code called focusModifier.requestFocus() before onFocusChanged but after onWindowFocusChanged. The existing code sets focusModifier state when we gain window focus. Which is incorrect. We need to wait for view focus too.

This CL removes the dispatchWindowFocusChanged callback because the default implementation calls onFocusChanged internally. If we use the default implementation we can just rely on the onFocusChanged callback.

Bug: 159480822
Bug: 159371598
Test: Verified that this fixes ProcessKeyInputTest and AndroidProcessKeyInputTests by running them multiple times.
Change-Id: Ie40996ab2872970a713d011ddf944e038a555641