Make it explicit that ui-animation-tooling lib is internal
Rename ui-animation-tooling to ui-animation-tooling-internal to make it
explicit it's meant to be used internally only.

The goal of the library is to provide APIs that will make it easier to
manipulate objects shared between ui-tooling and Android Studio. We don't
expect it to be used externally, so we mark it as Publish.SNAPSHOT_ONLY.

Bug: 159418525
Test: ./gradlew checkApi
Change-Id: Iccad455e652c36157be3cc47659e4a9579a8a68a