Add MailTo class to parse mailto: URIs (RFC 6068)
The MailTo class is meant as a drop-in replacement for
android.net.MailTo. The API is retained, but a couple of bugs present
in the platform version have been fixed.

See also: https://cketti.de/2020/06/22/android-net-mailto-is-broken/

Relnote: "Added `MailTo` API which provides consistent behavior and bug fixes for all API levels"
Test: ./gradlew :core:core:connectedCheck
Bug: 159827506
Change-Id: Ie93956270775e1214d778a890711105dd4b1db19