Fixed issue where unable to switch back to a camera.
This fixes the issue in CameraView where switching from rear to front
and then back to rear camera did not work. The issue is that during the
change to LifecycleCameraRepository the active camera would be the one
with the latest started LifecycleOwner. However, previously the
LifecycleCamera with the latest bound UseCase would be the active
camera. Reverted this behavior by adding a setActive() method to
LifecycleCameraRepository.

Bug: b/159848619, b/159836255
Test: LifecycleCameraRepositoryTest and test of view test app.
Change-Id: I5bcc071925728bc85a59435b9b4fcd6e46b56785