Fixed issue where Vectors would be aliased while drawn
Relnote: "Refactored PainterModifier
to no longer provide scaling based on
the given constraints and ContentScale
parameter. Implementations of Painter
are to scale their drawing content
themselves based on the given size of
the DrawScope they are drawing into.

Fixed issue where VectorPainter's cache
bitmap was sized to its default size
instead of the given size to draw into.

Updated ImagePainter to scale its
content instead of relying on PainterModifier
to do so on its behalf."

Change-Id: I6ba901d9622d5070c64ee641044d9bfddd409222
Fixes: 159619901
Test: Added tests to PainterModifier