Moved LifecycleCamera classes into camera-lifecycle
Moved LifecycleCamera and LifecycleCameraRepository classes.

Removed all of the CameraX.bindToLifecycle() related calls and replaced
them in the unit tests.

Remove check in extensions for active use cases so that LifecycleCamera
can be moved into camera-lifecycle.

Made sure that in LifecycleCamera, the attach/detach of the
CameraUseCaseAdapter matches that of the Lifecycle that was used to
create the LifecycleCamera.

Bug: b/147925044
Test: androidx.camera-{core,cameram2,lifecyce} tests
      manual test of view and core integration apps
Change-Id: I295138605dd0c73f4d2421160d9fc1274c510318