Remove `StableMutableList` and `StableMutableMap`
Temporarily remove `StableMutableList` and `StableMutableMap` to
avoid b/160018668.

Relnote: "Temporarily removed `StableMutableList` and `StableMutableMap`
to avoid an issue in the version of Kotlin compose requires. These
interfaces will be reintroduced once compose is updated to a version
of Kotlin that doesn't have the issue.

`SnapshotStateList` and `SnapshotStateMap` are now public but they will be
deprecated once `StableMutableList` and `StableMutableMap` are restored."

Fixes: b/160013170
Test: existing tests pass
Change-Id: Ia77691c6a5f523f63bd5d76cd6d7d7dc3f1afce1