feat: data transformations for 1D visualizations
Before we display data it is useful to perform some transformations on
it. For example, the lens focus distance metadata can technically
range from 0 to infinity, so it might be easier to visualize the log
value over time. We need to label these transformations as 1 dimensional
because for some data, there might be multiple transformations that make
sense. For example, given an array of faces, it might be useful to graph
the number of faces over time (1D visualization), but it might also be
useful to draw the corresponding face rectangles (2D visualization).
Thus in clarifying the data transformations as 1D  we are leaving room
for this possibility in the future.

Test: deterministic tests were performed on each 1D transformation defined
Change-Id: I65820939dcb4ad0d5c866b960d195cc0f85f2702