Fixes onPreferenceStartScreen being called twice
aosp/1343920 added support for checking getActivity() for this interface, as well as getContext(), but doesn't check to see if getContext() has handled the interface first. Now getActivity() is checked only if getContext() did not return true.

Also adds tests for the other interfaces that are working correctly, to ensure this does not regress in the future.

Bug: b/160238129
Test: PreferenceFragmentCompatInterfaceTest
Change-Id: If594047b0a8c7a7708d3153a9ba7a9be00a43d41