[RecyclerView] Avoid caching incorrect item decorator offsets when in pre-layout
**Background**
In our use case, we were using an `ItemDecoration` that applied item offsets. We saw a bug manifest in the form of incorrect item offsets appearing in the UI when reordering items. This is due to the fact that our offsetting is computed based upon span index.

**Test Case**
In this change, I added a test case that...
- Reorders items
- Asserts span indices within the decorator

Prior to the change in RecyclerView, it would fail by calling the decoration with the span indices not yet updated. This wouldn't be an issue if it simply invoked again (it would just be an inconsistency), except that RecyclerView will cache the bad value on the first pass. As a result, the bad value gets used going forward resulting in potentially broken UI.

I think there's probably room to have some sort of test just for RecyclerView, since my favorite approach ended up being in RecyclerView itself after initially resolving in SGLM itself, but I think this may be fine. Let me know if you disagree!

**Change**
Initially, I simply added a short-circuit in `StaggeredGridLayoutManager` so that if we were in `preLayout`, we didn't measure. Works fine, but I spent some time staring at the underlying caching logic in RecyclerView, and it has some logic there as well seemingly designed to avoid caching the wrong stuff. This felt like a logical extension: if you're going to move an item and you're in pre-layout, don't bother computing things yet. Some debug logs that help show what's going on:
```
Calculating item decorations: TestViewHolder{f2d6e9 position=3 id=-1, oldPos=4, pLpos:4} item:Item{mId=5, originalIndex=3, text='Item '}, data:null
Old position: 4
Layout position: 4
Absolute adapter position: 3
Binding adapter position: 3
Is pre-layout: true
```

At this point, the span index on the LayoutParams will be (not) updated to reflect the layout position, resulting in those incorrect values being cached as the decor insets.

Bug:37140917

Test: testItemDecorationLayoutParamSpanIndices() validated that item decorations were computed with correct values.

Change-Id: I2898947a84f565cda9910bd945b74bc3f88d60d9