Add accessibility action to get TextLayoutResult
1. SwitchAccess/VoiceAccess is still relying on framework for
granularity movements, so it is safer to implement the traversal action in
accessibility delegate so that every semantics node has this action.
2. Compared to text/textfield implement the traversal action, providing
TextLayoutResult to accessibility can avoid duplicate code (traversal
implemention in both accessibility delegate and text/textfield.). And we
avoid exposing the traversal action to compose developers.
3. For text location API, we need to get a list of Rects for the specified
text locations. We can use the added property in this CL or add a new
action to passback a list of rects. I prefer adding fewer APIs.

Relnote: Add accessibility action to get TextLayoutResult

Test: tested with later code of granularity traversal.
Change-Id: I9d6e6313528500524f04638ccb5742fcfbb41392