Fixed SupportedSurfaceCombination aspect ratio size grouping issue
Group all available sizes by aspect ratio first and then sort the groups by the distance to the target aspect ratio to find the best suitable size.

The developers may set the target resolution as any arbitrary size. Original logic directly used the aspect ratio of the target resolution setting to find the supported sizes that match the aspect ratio under mod16 mechanism. It made some supported sizes that should belong to other aspect ratio were incorrectly grouped under mod16 mechanism, and then considered to match the aspect ratio of the target resolution setting. For example, 864x480 should be a 16:9 size, but it was grouped to match the aspect ratio of 2016x1080 target resolution setting. To fix this issue, the all supported sizes will be grouped first. It can avoid a size to be grouped to match a strange aspect ratio. Then, the auto-resolution mechanism will sort all sizes by the aspect ratio difference distance of the groups to select the suitable supported size for the target resolution setting.

Relnote:"Fixed auto-resolution aspect ratio size grouping issue that a 16:9 mod16 size (864x480) is selected when the target resolution setting is 2016x1080 and there is another 1920x1080 16:9 size supported."

Bug: 159363774
Test: run SupportedSurfaceCombinationTest
Change-Id: I531671972a05bf01e2dea1f96da5bc3d6655c391