Remove type inference in transitionDefinition
Type inferece in 1.4 seems to have trouble with
the overloaded transition functions in TransitionSpec,
and inferred the overly-generalized type:
e.g. Pair<T,T> is inferred as Serializable

With that said, it's more desireble to require
explicity, in order to avoid mysterious (and silent)
failures.

Relnote: "Require type T to be explicitly specified
for transitionDefinition."

BUG: 149673850
Test: existing tests pass

Change-Id: I1adedb34525ebb8c079a77a9af2636f1cb8339f7