Clear existing link on new one in ConstraintLayout
Before this CL, when adding a new link in ConstraintLayout, the
potentially existing links from the same origin were not cleared. Hence,
when resolving the constraints, the solver could see, for example, both
LEFT_TO_LEFT and LEFT_TO_RIGHT anchors. The winning one would be
arbitrarily resolved depending on the ordering they were considered in
the solver code. This CL prevents this, by clearing the possibly
existing link from the same origin, when a new link is set.

Fixes: 161217731
Test: ConstraintLayoutTest
Change-Id: I61f46c147fc0cb5bcc65fa76b21438765e25edc2