Bugfix: flip captured image if camera is front camera.
Introduced in aosp/1285103, CameraView no longer sets the flip flag in metadata.

Fix: overwrite the setReversedHorizontal flag. Alternatively, we can call setReversedHorizontal only when it's not set. But it requires us knowing if setReversedHorizontal() has been set before (and a way to clear it if set), which complicates the logic. Always overwriting is simpler.

Relnote: for image capture, overwrite the flip horizontal flag in metadata based on camera direction.

Bug: 161802053
Test: manual test and ./gradlew bOS
Change-Id: I28499bd3a46b0760d31fe100f40175552e6b6221