Fix a flaky TokenLock test.
The test flakes indicated that the TokenLock was not handling suspend
calls when the the continuation for an acquired token was canceled.

This only way (within the context of this test) for the TokenLock to
return "0" in this situation would is if there is a pending request in
the queue *after* the canceleation has completed. Further inspection of
the `invokeOnCancellation` method indicated that if this method was
invoked synchronously the pending request would be added to the queue
after the cancellation callback ran. This would cause the request to
remain in the queue, and causing the TokenLock to report 0.

Fix: Register the cancellationCallback after the request is added to the
     queue so that it gets processed during cancelation.
Fixes: b/161897382
Test: Locally increased the number of iterations to 10000 for this test.
Test: Existing tests continue to pass.
Change-Id: I2920aa74ad7521a73a621e3975c5e3d8ff7a477f